+++ library/std/src/sync/once.rs
