+++ library/std/src/sys/sync/once/no_threads.rs
@@ -1,6 +1,6 @@
 use crate::cell::Cell;
 use crate::sync as public;
-use crate::sync::poison::once::OnceExclusiveState;
+use crate::sync::once::OnceExclusiveState;
 
 pub struct Once {
     state: Cell<State>,
